12_Google_Vision/1_Vision_API/requirements.txt
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud==0.34.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
libcst==0.3.15
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
